requirements.txt
fastapi
uvicorn
streamlit
langchain
openai
faiss-cpu
python-dotenv
pdfplumber
python-docx
scikit-learn